requirements.txt
django-cors-headers==3.1.1
swagger-spec-validator==2.7.3
Django==3.1.7
flake8==3.8.4
jupyter==1.0.0
django-extensions==3.1.1
django-ckeditor==6.0.0
requests==2.25.1
djangorestframework==3.12.2
drf-extensions==0.7.0
django-heroku==0.3.1
gunicorn==20.0.4
django-filter==2.4.0
drf-yasg==1.20.0
whitenoise==5.2.0
djangorestframework-simplejwt==4.6.0